+++ lib/loaders/network_file_translation_loader.dart
@@ -22,6 +22,7 @@ class NetworkFileTranslationLoader extends FileTranslationLoader {
     networkAssetBundle = NetworkAssetBundle(baseUri);
   }
 
+  @override
   Future<String> loadString(final String fileName, final String extension) {
     return networkAssetBundle.loadString('$fileName.$extension');
   }